requirements.txt
Flask==3.0.2
Authlib==1.3.0
SQLAlchemy==2.0.29
Flask-SQLAlchemy==3.1.1
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
